requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.22.11
botocore==1.25.11
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-appconf==1.0.5
django-cleanup==6.0.0
django-cors-headers==3.12.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-decorator-include==3.0
django-filter==21.1
django-redis==5.2.0
django-select2==7.10.0
django-storages==1.12.3
django-tables2==2.4.1
idna==3.3
jmespath==1.0.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.1.0
wrapt==1.14.1
